requirements/requirements-project.txt
django-ckeditor==5.9.0
django-vinaigrette >= 2.0, <= 2.1
djangorestframework >= 3.11, <= 3.12
djangorestframework-csv >= 2.1.0, < 2.2.0
django-haystack == 2.8.1
whoosh >= 2.7.4, < 2.8.0
django-jet==1.0.8
vcrpy
drf-nested-routers
requests
django-resized==0.3.9
pillow
shortuuid
djangorestframework-jwt
dateutils
pyexcel_xls
django-cors-headers
django-rest-framework-social-oauth2
martor
django-admin-rangefilter
boto3
beautifulsoup4
python-dotenv
gql
django-import-export==2.0.2
openpyxl
tablib==1.1.0
celery
simple-salesforce
